Feat / Separating Componnent(AnimeCard) (#6)
